improved actions' permissions

+++ actions/action_delete_ticket.php
@@ -1,13 +1,30 @@
 <?php
-include_once('../utils/session.php');
-include_once('../database/connection.db.php');
+require_once('../utils/session.php');
+require_once('../database/connection.db.php');
+require_once('../database/ticket.class.php');
 
-// ver se user está logged, se o ticket é dele ou se é agente ou admin
+$session = new Session();
+
+if (!$session->isLoggedIn()) {
+    header('Location: ../pages/login_page.php');
+    die();
+}
+
+$userType = $session->getUser()->type;
+$isAdminOrAgent = ($userType === 'admin' || $userType === 'agent');
+
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+
+    $db = getDatabaseConnection();
 
     $ticket_id = $_POST['id'];
 
-    $db = getDatabaseConnection();
+    $ticket = Ticket::getTicket($db, $ticket_id);
+
+    if (!$isAdminOrAgent && $ticket->ticketCreator->id !== $session->getUser()->id) {
+        header('Location: ../pages/userTicket.php');
+        die();
+    }
 
     $stmt = $db->prepare('DELETE FROM Tickets WHERE id = ?');
     $stmt->execute([$ticket_id]);
